Replace laser by new_laser
New laser is an animated sprite with a fixed circle collision shape on
the tip.
Laser velocity is fixed to 600 pixel per second and so the animation of
the sprite follows that speed.
The previous laser was giving problems with the collision shape and
also when the game is scaled by window it is odd that the laser wasn’t
scaled too.